requirements.txt
annotated-types==0.6.0
Authlib==1.3.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.6
install==1.3.5
protobuf==4.25.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.1.0
validators==0.22.0
weaviate-client==4.4b0
